pyproject.toml
[project]
name = "vibeout"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "mysql-connector-python (>=9.4.0,<10.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
